+++ Cpp/PracticeProject/BankProject/DonationAccount.h
@@ -1,15 +1,15 @@
 #pragma once
-#include "Account.h"
+
 class DonationAccount : public Account
 {
 public:
 	DonationAccount();
 	~DonationAccount();
-	virtual void Deposit(float money) override;
+	virtual void Deposit(int money) override;
 	virtual void Withdraw(int money) override;
 	virtual void Print() override;
 private:
-	float donation;
+	int donation;
 	const float donationRate = 0.01f;
 };
